Add CancellationToken to the RemoteLockHandle to detect loosing the lock

+++ Cassandra.DistributedLock/TimeSpanExtensions.cs
@@ -4,9 +4,9 @@ namespace SkbKontur.Cassandra.DistributedLock
 {
     internal static class TimeSpanExtensions
     {
-        public static TimeSpan Multiply(this TimeSpan value, int factor)
+        public static TimeSpan Multiply(this TimeSpan value, double factor)
         {
-            return TimeSpan.FromTicks(value.Ticks * factor);
+            return TimeSpan.FromTicks((long)(value.Ticks * factor));
         }
     }
 }
